[bind] Remove unconditional bind in new bytecode
The compiler currently inserts Unconditional Bind
at the end of the bind rules. This is an issue for
the new bytecode since it does not support match
instructions.

To fix this issue, the CL adds a bytecode version
field in BindProgram. The compiler will then
insert Unconditional Bind based on the bytecode
version.

Bug: 73154
Tests: Updated existing tests. Added test in
       compiler.rs to check.
Change-Id: I052fae89a301f52d15005747e1783f278504a2d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508208
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>