[packet] Impl ExactSizeIterator for RecordsIter
When constructing a `Records` object, we perform a validation pass over
the records. This allows us to count the records during this pass and
know for future iterations how many records will be produced, which in
turn allows us to implement `ExactSizeIterator` for `RecordsIter`.

Test: fx test packet-test -- all_records_parsing

Change-Id: I2ebcf78b2c52f90c97865163364e4ab1fba6e3ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507139
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Arthur Sfez <asfez@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>