[fidlgen][cpp] Helper for doc comments
Basically, replace:
  {{- range .DocComments }}
    ///{{ . }}
  {{- end }}
with:
  {{- .Docs }}
in C++ bindings.

This makes the templates have less visual and semantic noise. Perhaps
most importantly it removes a scope so there's one less {{ end }} to
keep track of.

To do this we have to be consistent on how we put doc comments into C++
and I went with triple-slash as was used in the HLCPP comments because
it's supported by Doxygen.

Change-Id: Ia567e16adebaf64c8b1c9e19aecb7d7dd5ee885a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510083
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>