[ffx] Correctly get tools from modular SDK in tree
Moving to the modular SDK will put the in-tree manifest in a new
arch-specific location. The upside of that will be only one of each tool
in the manifest, and selecting the right manifest means always getting
the tool with the right arch.

Right now building the modular manifest is optional, but to be ready for
the transition, we always look for it and fall back to the old location
if it's missing.

Change-Id: Id58baf4d6d93c3e401048d829878875ed0c5e507
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506865
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>