[zircon][kernel] Support for ZX_INFO_VMO_IMMUTABLE
This CL is an implementation of RFC-0080 and adds support for
determining if a VMO child was created in an immutable state (and will
remain immutable).

For this to be the case, the VMO clone must be created with the
ZX_VMO_CHILD_NO_WRITE and ZX_VMO_CHILD_SNAPSHOT flags and without the
ZX_VMO_CHILD_RESIZABLE flag. The child VMO also must not be discardable,
but there currently is no way to create a discardable child (and
therefore also no way to test for this case).

Change-Id: I16c1980ee82428a3184d4f72643e57fa0251aaf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508658
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>