[session] Add scene manager component
This component extracts scene management and input code out of the
workstation session. This first step keeps all the functionality as
similar as possible, and introduces a simple FIDL protocol that
workstation can speak to the scene manager to configure the scene and
input pipeline.

This FIDL protocol will need to be updated prior to inclusion in the
SDK, as it currently contains a legacy input event listener.

Ideally this component will, after removing some of the workstation
specific quirks, be able to be used across products to set up the scene
graph and handle input.

Change-Id: I79689209adf18396dff75c14517d4fc36b5f0753
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487757
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
API-Review: Chase Latta <chaselatta@google.com>