[testsharder] Run multiplied test at least once.
There was a bug where testsharder would calculate the number of runs for
a test (if the total_runs was unset) by dividing the target shard
duration by the test duration. If the target duration was less than the
test duration, than the number of runs calculated was 0 so it would
assign 0 tests for that multiplied shard to run. This was written to the
json output file as `null` which would give a type error when processed
by the recipes. We should instead run it once.

Bug: 73515
Change-Id: I88fdbab6921db87f611dbb6e0ea8e06e130b2ab2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510087
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>