[kernel][asan][arm64] Use new ArchAspace::Map interface.
The new interface for mapping pages requires the caller to specify an
action to perform when a mapping already exists in place.

We are using the same behavior as before: fail. The code in kasan
already checks whether a mapping exists before doing it.

Change-Id: Ic8914788e06a3d7fbf9329be3d2cf828ede8fa59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510321
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>