[bt][at] Add support for optional arg delimiters.
Typically, AT command arguments are offset from the command name with an
=.  Sometimes, however, they can be offset with > or no character at
all.

The first of these cases was already supported as a "nonstandard
delimter" in the AT command library.  Adding support for the second
required some restructuring of the Definition and Command types.  The
optional nonstandard delimiter is changed in this commit to an optional
delimiter string.

This introduced some issues in parsing, as the parser can now see the
lack of arguments to a command as an empty delimiter, followed by an
empty argument list.  To remove this ambiguity, I made execute commands
arguments mandatory.

This necessitates changes to tests and generated translation code.

Bug: fxbug.dev/66041 fxbug.dev/72873
Change-Id: Id92f8b8ad1c19aa36bd63e48aa110c9a6a910546
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505561
Commit-Queue: Jeff Belgum <belgum@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>