[fidlc] Implement parsing of layout subtypes.
Value layouts `enum` and `bits` previously did not support the parsing
of "wrapped type" in the new syntax.  The code to parse these tokens is
a bit tricky, as the parser cannot unambiguously decide if a colon that
appears like `enum : FOO` denotes the start of a constraints block, or a
wrapped type.  This is resolved by parsing `FOO` as a constant, then
checking the next token to see if its a `{` - if it is, we assume that
`FOO` is a wrapped type, otherwise, we assume it is a constraint.

Test: fidldev test fidlc
Bug: 65978
Change-Id: I4e9fc6f4c7154c0048f9941000990405b559c587
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509805
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>