[zbitl] Abstract ZBI-specific logic in View
Soon we will move the core of View's container-y business logic into a
subclass templated on "container traits" (which will too will be used
for a BOOTFS analogue). To ease that transition, we introduce the
types, and static members and methods that will comprise the traits of
the ZBI format.

`header_type` was renamed to `item_header_wrapper` to avoid confusion
with `item_header_type` and `container_header_type`, both introduced in
this change.

Container header logic is updated to no longer takes advantage of the
optimization provided by header_type/item_header_wrapper, as BOOTFS
container and item headers are distinct. The cost of this, however, is
negligible: it amounts to at most a handful of bytes copied over the
lifetime of a View.

Bug: 68585
Change-Id: I5210f31a88c4df8b9f37a0ffdffbccb586097170
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490297
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>