[carnelian] Don’t render when display isn’t owned
The main problem with this approach is that when ownership of the
display is returned, an incorrect pre-rendered image is shown. I’d like
to go with this now, though, and take more time to consider how to make
sure the first frame on regaining ownership is correct.

I noticed that the fuchsia-framebuffer example box was a little broken
so I fixed it and updated the manifest file for the integration test.

Bug: 73628

Testing: manual, by checking that spinning square doesn’t appear in top
when it doesn’t own the display.

Change-Id: Id473f2f8b85b1ba1082298b474dbf52b16125315
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510582
Commit-Queue: Rob Tsuk <robtsuk@google.com>
Reviewed-by: David Reveman <reveman@google.com>