[selectors][archivist] Support recursive glob
Adds support for a recursive glob, which allows a selector to specify
all descendants of a node. Recursive globs are only allowed as the
final segment of a component selector.
This change also explicitly disallows use of recursive globs in static
selectors so that a change in component topology does not unintentially
increase the scope of exported diagnostics.
This does not add support for rersive glob to remote-control.

Bug: 68912
Testing: new integ test cases to verify Archivist supports new
selectors, new unit tests for selector lib.
Multiply: archivist-integration-tests-v2

Change-Id: I5479caf402e1c4314bedcddfe9d8bab66fe7f642
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507398
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Luke Nicholson <lukenicholson@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>