[camera] Handle client token failures async
Instead of using ValidateBufferCollectionToken to see if a token is
valid, register an error handler and call Sync. This eliminates the need
for the client to call Sync before using SetBufferCollection and also
improves handling for the error case of a token server end being closed.

Also ensure that a client provided token is not included in other client
buffer requests until a Sync has completed to validate there's a
responder on the other end.

Adjust unit test to test for the case of a token failing after being
passed to SetBufferCollection.

Bug: crbug/1192720
Test: camera_device_unittest
Multiply: camera_device_unittest
Change-Id: I36427099eacc294d55dd708c7fe6670f80774102
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509044
Reviewed-by: Ernest Hua <ernesthua@google.com>
Commit-Queue: Austin Foxley <afoxley@google.com>