[covargs] Run llvm-cov on individual binaries to check for validity.
Some of the Rust test binaries are giving a `Malformed instrumentation
profile data` error. This should be looked into but in the meantime we
should skip these files so that we can still generated the coverage
report with the rest of the valid profile data.

Bug: 27171
Change-Id: I21b3fe649f0db45262c74f25e8a9c6f97ba0211f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511398
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>