[fint] Generate ninja graph and compdb
Currently this is done in recipes, but once `fint build` is done we'd
like for the recipes to no longer know how to run Ninja at all. So it
makes sense for `fint build` to handle these.

We pass the paths to the generated files via output build artifacts
message so that the recipe doesn't need to hardcode constant paths to
these files.

Also emit the path to the main ninja log file, as the infrastructure
needs access to that file to do some performance result processing. The
path is constant with respect to the build directory so theoretically it
could be hardcoded in the recipes, but that path is more of an
implementation detail of the build system that falls solidly on the
platform side of the platform-infra interface, so it's best to not
hardcode that sort of knowledge in the recipes.

Bug: 67861
Change-Id: Iebf0e409437ad014be814044790a1f1a256fa0fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511441
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>