[kernel][oom] Unify sync and async eviction paths
Use a common wrapper for synchronous eviction (which is used at OOM) and
asynchronous eviction (which is used at Warning and Critical), to reduce
the likelihood of introducing bugs in divergent code paths.

Add a regression test that verifies pages are evicted out of a pager
backed VMO on forced reclamation, and that free memory count increases.

Test: core-pager

Bug: 73865
Change-Id: I7d1eabbcbfa035e49336f00a2952c73eae523932
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511678
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>