[build] Add template for verifying two files match
Add a new GN template and python script that tests at build time if
two files have the same contents.

This is a utility for verifying that migrations are resulting in the
same contents by allowing an old and new implementation to reside
side-by-side in the tree, and verifying that they are producing the
same results.

This is not meant for "Golden File" verification, but is mechanically
very similar.  This is for comparing two build outputs against one
another.

Bug: 73184
Change-Id: I069ab171c1d983594fe11e8c044c397a2881feee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511638
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Shai Barack <shayba@google.com>