[driver-manager] Seperate Matching from Binding
In coordinator we should seperate 'Matching' a Device's
properties to a Driver's Bind code to the act of 'Binding' the
Driver to the actual device.

Once this is seperated, we will be able to keep all of the
Drivers in a seperate DriverMatcher class, which will eventually
be replaced by the DriverIndex component.

Bug: 73827

Change-Id: I75f6e1a61cf704863b051271fbea15f440f2fc12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510800
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>