[scenic] Simplify FrameScheduler initialization
Removes AddSessionUpdater() and SetFrameRenderer() from the
FrameScheduler interface and instead creates a single Initialize()
method for DefaultFrameScheduler where they're all set at once.
Ideally this would be part of the constructor, but due to the circular
dependencies this creates that's currently unfeasible.

Adds a guarantee that SessionUpdaters will be called in the
order they're submitted to DefaultFrameScheduler. This is necessary for
integrating ViewTreeSnapshotter with the FrameScheduler.

As part of cleanup I also merge session_handler_test directly into
vk_session_handler_test since it's the only remaining dependent.

Bug: 73451
Test: fx test gfx_unittests input_unittests scheduling_unittests

Change-Id: Ib084e6d13be3d387794f1daff09d47211497b745
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509472
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>
Reviewed-by: Felipe Archondo <felipearchondo@google.com>