[netemul] Use test driver pattern
This will allow us to add components such as network_context and
sync_manager as *siblings* of the netemul_sandbox component, as opposed
to being children of it, so that we can route capabilities between them;
for example, the sandbox needs to be able to use
fuchsia.netemul.network/NetworkContext.

This CL also updates the inspect test to expect the 'test_driver' prefix
in the moniker of the components under test.

Bug: 72402, 72403
Test: fx test -d netemul-v2-tests
Change-Id: Ifda06138be374764b09a2fa0709c4b462ac28526
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510759
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>