[logs] Update text for logs that were rolled out
Currently, we say that logs which were rolled out of the
buffer are "dropped". This is confusing, because it
conflicts with logs which were dropped at the socket level.

This patch updates the display to use the term "rolled out" instead
of "dropped" for these logs.

Fixed: 74001

Change-Id: I80e75d137045b7673a20979dbf5a299cd50039ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512035
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>