[build] Fix debug archive listing for prebuilt_package_with_flavors()
This CL fixes the prebuilt_package_with_flavors() template to properly
pass the flavor's "debug_archive" field to prebuilt_package().

This is necessary to ensure that debug archives are properly unpacked
and their content listed in prebuilt_binaries.json, which is now
required to upload the corresponding debug symbols to the symbol server.

For more details, see the associated bug.

The issue was fixed in ToT by [1], which also modifies how prebuilts
for chromium are handled by the build. This CL only replicates the
part of this change that modifies //build/packages/... instead, since
a simple cherry-pick to the F1 branch may lead to other unwanted
changes. This will also simplify future merges of ToT into F1.

For testing, do the following:

  # Configure the build. There is no need for extra bundles.
  # Replace VENDOR_PRODUCT and VENDOR_BOARD with appropriate values
  # (see bug for details).
  fx set VENDOR_PRODUCT.VENDOR_BOARD

  # Verify cast_agent debug symbols are listed, there is no need to build!
  grep "cast_agent" out/default/prebuilt_binaries.json

Before this CL is applied, the output was empty since there was
no entry for "cast_agent" in the prebuilt_binaries.json file. After the CL is
applied, it should contain something that looks like:

  {
    "debug_archive": "../../prebuilt/build_ids/arm64/castagent/cast_agent.symbols.tar.bz2",
    "manifest": "gen/vendor/google/bin/cast_agent/cast_agent.debug-archive_binaries.json",
    "name": "cast_agent.debug-archive"
  },

So the grep will return the three lines above.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/484577

Bug: 73971
Change-Id: I2b554ed5cd22a6c2f982a7297204e0e44bf8969d