[debugger] Detect languages from file names.
When the user says "list <filename>" or "list <line>" there was no
symbol to take the expression language from and the list output was not
highlighted.

This change adds file name extension checking for when no symbol
language is known, and falls back to C highlighting for everything else.

Change-Id: I78855f0407b4311bc6336f215e998063f3d27e7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511747
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>