[fvm][tool]: Remove legacy create.
Add small missing bits, and edge cases as I migrated the create command
to use the volume image library.

Fixed trim method and added test. It was passing a pslice 0, and not
taking into account the end of the allocated slice.

Fixed blobfs label, being set to blob instead of blobfs. This caused
blob partition to be ignored.

Make the slice size explicit in the image tests.

Test: CQ
Change-Id: Ide76b0f42f7d4b22fc231bc3eac3d5cf4f165f7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510579
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>