[carnelian] Enable the use of the virtcon display mode in carnelian.
This makes it possible for fuchsia-framebuffer to bind to
a virtcon display client, so that carnelian apps could
run as virtcon.

I took out the build changes related to actually running as virtcon
to make this CL, based on Yilong’s work, just modify Carnelian
and fuchsia-framebuffer.

Change-Id: I4b8d7ab4a0185094f9dd1d20230e5302b4f11c9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511619
Commit-Queue: Rob Tsuk <robtsuk@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: David Reveman <reveman@google.com>