[graphics] Convert vulkan_loader to components v2
This is the first step in loading Vulkan ICDs from packages. After this
we can modify the loader to launch V2 components that will provide the
ICD.

Also add the vulkan_loader package to the base packages in core.gni to
ensure that it will be found properly - it seems like universe packages
aren't resolved.

Bug: 13078

Change-Id: Ib2f403b7e920dc431a346d8ad8b336f3275b016a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487139
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: John Bauman <jbauman@google.com>