[fx build] Add --fint-params-path flag
For consistency with `fx set`. If set, `fx build` will bootstrap fint
using the same bootstrap.sh script as infra, and then run `fint build`
with the specified fint params file and a context spec textproto file
constructed at runtime.

Since Bash doesn't have a protobuf library we need to construct the
context textproto as a raw string. That's not great since there's no
schema validation, but we only need to set two top-level string fields,
so it's not really worth writing a separate Go tool just to do that.

Eventually `fx repro` will set the --fint-params-path flag in the `fx
build` command that it produces, making it possible to exactly reproduce
the set of targets built by a given builder.

Also, make some adjustments to `fint build` so that it will produce the
same in-line logging behavior as `fx build` when run locally, instead of
printing each line of output to a new line and filling up the terminal
with potentially hundreds of thousands of lines.

Bug: 67861
Change-Id: I47a28662e5bc3891301ec49628c925e6bd33ce65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512598
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>