[build][ffx] Add GN template for calling ffx from within the build
Add a GN template to correctly call the compiled_action() template
to use ffx as a compiled build action.

This template is a thin wrapper around compiled_action(), and will
automatically add the dependency on the correct ffx build label,
and the binary to the inputs.

The template parses a string of the "plugin name" into individual
tokens for passing to ffx, as well as properly formats a key=value
list of config overrides.

Bug: 73185
Test: local tests using the template
Change-Id: I1e1b3d2e298255fa9ff353bde2e9cd0921bff195
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511679
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: James Tucker <raggi@google.com>