[fx] Evaluate $_FX_LOCK_FILE greedily
Instead of evaluating it at the time the `trap` executes. This ensures
that the original file gets cleaned up even if the value of
$_FX_LOCK_FILE is changed for some reason (which should currently never
happen, but we might as well be resilient to that).

This undoes a small portion of fxrev.dev/512098.

Change-Id: I9b5cb14b01f10ad7d1293817e9d2dcab927842c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513140
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>