[kernel][vm] Eliminate warning while reserving range for periphmap
The call to ReserveSpace in reserve_periph_ranges was failing with
ZX_ERR_INVALID_ARGS and generating a "conflicting cache policies"
warning.  Why was it failing with ZX_ERR_INVALID_ARGS?  Because
SetMappingCachePolicy was failing, because LookupLocked was failing,
because it is an error to call LookupLocked on a zero-sized VMO.

In SetMappingCachePolicy, only call LookupLocked (and
arch_clean_invalidate_cache_range) if the VMO has a non-zero size.

In reserve_periph_ranges, ASSERT that the call to Reserve succeeds.

Fixed: 74086
Change-Id: Idc6650895e38f69d4c6bc8df6ab40970c81fd7e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511633
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>