[tefmocheck] Don't attribute logs to tests by default
Many logs are never caused by a specific test, but rather by random
underlying flakiness in the operating system or the infrastructure. It's
not useful to attribute these logs to the tests during which they
happened because there's no pattern to it and it just leads to an
excessive number of flake bugs being filed, all with the same root
cause: for example, 87 flake bugs are currently open for the log
signature ".sock: write: broken pipe", each bug specific to a test:
https://bugs.fuchsia.dev/p/fuchsia/issues/list?q=_broken_pipe%20component%3AFlake&can=2

So only attribute a few specific checks to the test during which the log
occurs.

Bug: 67640
Change-Id: I1987edf38ee0b9c05eb60329e4ea0491ef8c82b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513220
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>