[bt][rfcomm] Fix L2CAP connect not completing
When connecting through RFCOMM, if the FIDL client has to wait for the
response, it registers a waker with fidl::Client.  Because of how the
fidl::Client manages simultaneous two-way requests, all of the
outstanding requests must be polled to guarantee waking any particular
request.

RFCOMM did not do this, as some requests (advertise responses) were
polled in a select which is waiting to be awoken from
Profile.Connect() requests.

Instead, spawn advertisement requests on their own task, which will be
polled whenever woken.

Fixed: 73267
Test: added test_connect_l2cap_channels_while_advertising
      fx test bt-rfcomm-tests
Change-Id: Iad6a3a0080be31ba3d901760a7ff21ec4a4958be