[volume_image]: Introduce BlockWriter.
There is a custom use case, when paving a device on migration that
writes directly to a block device. This requires aligned writes.

To simplify this, in practice all writes are aligned, or mostly, this
will readback the unaligned blocks, update the the blocks and write them
back. The aligned blocks are passed through.

This should replace the mtd-file in the fvm tool.

Test: storage-volume-image-tests
Change-Id: I291e295bbd0c6af9bf93608fc60349c05d5cfb2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512306
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>