[pkg-cache-integration-tests] put isolated devmgr in package
Previously, the tests were failing locally because the devmgr we were
trying to use was in a different package (storage-isolated-devmgr).

I suspect this was working in CQ because that other package was available...
which may not be true locally. Either way, let's fix it to work locally.

Test: fx test -o pkg-cache-integration-tests
Change-Id: I05ca8ec1a7c5be7b16aee21b13943ff2e16c22fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513297
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>