[netstack] Cleanup expiring DNS server timer
Previously, when an interface was brought down, we did not stop the
timer for expiring DNS servers. When the interface came back up and
rediscovered an expiring DNS server, the server may be prematurely
expired by the stale timer that was not stopped when the interface was
brought down.

Test: fx test netstack-dns-gotests -- -test.run TestTimerCleanup

Fixed: 74201

Change-Id: Iee1db80c160a4ec102f05960ef2bb901e97aca66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513329
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>