[fint] Support incremental builds
This is logic that currently lives in recipes:
https://fuchsia.googlesource.com/infra/recipes/+/e963bc1e4f3d69b73c55464c5ea1ea5fcc221bf9/recipe_modules/build/api.py#719

However, now that fint is the status quo we're trying to adding
hardcoded paths to checked-in files in the recipes codebase,
particularly for build-related files in the main fuchsia.py code path,
as they make recipes more complex and force soft transitions between
recipes and fuchsia.git.

So add a context spec field for the recipe to indicate to fint that it's
building incrementally, and have fint handle updating the mtime of the
file. After this lands I'll update the recipes to set the `incremental`
context spec field and then delete the "touch" step from the recipes.

Bug: 64127
Change-Id: I9027ca315478163cfd485f7ddd8777c537043c4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513800
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>