[blobfs] Make host tool allocate enough inodes
The host tool was not creating a large enough initial inode table if
there were more files than kBlobfsDefaultInodeCount. Fixed this to
create an inode table that is just big enough.

The resultant image would be almost completely at capacity, so it won't
be very useful (although generally we don't use raw blobfs images, we
stitch them into an FVM image, during which we resize the inode table,
so this is most likely moot and easily remedied at that later point).

This change revealed many places (mostly tests) where assumptions about
the inode table size were made, so a number of corrections are also
included in this CL. Comments were added for some of the more subtle
situations.

Online formatting was refactored in this change to fix a number of
issues that I found while fixing the original problem. Namely, in the
FVM case, we do not ask the FVM device for the number of blocks in the
partition and use that value to set up the superblock; FVM devices only
report what is allocated, so this initial value is meaningless and would
only tell us what the previously formatted partition had allocated.
Ignore this value and simply proceed with attempting to allocate enough
slices for a minimal image (which is exactly what we did before,
anyways).

Bug: 73829
Change-Id: I9c6d32dbd7fdec6e5c2885057c09321af2290066
Tested: Unit.
Tested: Manually tested by generating a large corpus of files.
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512921
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>