[blobfs] Separate vmo from mapper in Blob.
Separate out the vmo from the data mapper in the BlobLoader. This is
required because the new pager uses a vmo managed by the PagedVnode base
class and has a separate mapper. This should have no behavior change.

This requires changing the blob loader to separate out the mapping,
buffer, and vmo in various places.

Add FreeVmo() function to Blob in the old pager case for source compat
with the new pager.

Change-Id: Ie58917c9e19c7a31b842e13706e606c9f2f9d6ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511918
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>