[wlan][policy] add lookup_compatible
This change adds a function to lookup all configs that can be used to
connect to a network with the provided detailed security type (as
opposed to the lookup function that uses a simplified policy security
type). The configs that are compatible are the ones whos security type
can be upgraded to match with the provided security type, ie we upgrade
wpa2 to wpa3 so if we lookup compatible configs for WPA3 we could return
one saved as WPA2.

Bug: 73578
Test: added unit tests
	fx test wlancfg-tests

Change-Id: I161087e1e54b2a370d65a5124f640362424b57b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512029
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>