[logs] Remove arc-swap dependency from Rust backend.
There's a newer version we're looking at updating to, but it has a lot
more unsafe and will be very high effort to verify. It's only used here,
so we're going to be a bit risk averse and just remove it from the tree.

AFAIK there's nothing that will suffer seriously when going from an
ArcSwap<Severity> to a RwLock<Severity>, and the .read() calls might
even be faster than ArcSwap's fast path when uncontended.

A future change will actually remove the crate, as its still depended on
by an out of date version of signal-hook-registry which will we'll
update past in fxr/512238.

Change-Id: I048d5750d730b352e811c2a4ebe3d34db6bc29fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514600
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>