[fx] Correct the docs and usage of FUCSHIA_DIR
Correct the documentation for FUCSHIA_DIR to note that the
environmental variable is not used.

Clarify the output message when the Fuchsia root is not found
above the current directory.

Update other copies of the root directory search code for
consistency with the recent deprecation of the FUCSHIA_DIR
environmental variable.

Change-Id: Id7ce316b70e7162ed1d75f17e5b9023e773c96a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513296
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Gary Boone <gboone@google.com>